+++ src/main/java/com/dev/delta/controllers/BreakdownController.java
@@ -18,7 +18,7 @@ import com.dev.delta.entities.Breakdown;
 import com.dev.delta.services.BreakdownService;
 
 @Controller
-public class BreakdownController  {
+public class BreakdownController {
 
 	/**
 	 * breakdownService
@@ -26,22 +26,22 @@ public class BreakdownController  {
 	@Autowired
 	private BreakdownService breakdownService;
 
-	
-
 	/**
 	 * getBreakdowns
+	 * 
 	 * @param model
 	 * @return
 	 */
 	@GetMapping("/breakdowns")
 	public String getBreakdowns(Model model) {
 		List<Breakdown> breakdowns = breakdownService.getBreakDowns();
-		model.addAttribute("breakdowns", breakdowns);	
+		model.addAttribute("breakdowns", breakdowns);
 		return "breakdown/breakdowns";
 	}
 
 	/**
 	 * addBreakdown
+	 * 
 	 * @param breakdown
 	 * @return
 	 */
@@ -54,6 +54,7 @@ public class BreakdownController  {
 
 	/**
 	 * findById
+	 * 
 	 * @param id
 	 * @param model
 	 * @return
@@ -67,6 +68,7 @@ public class BreakdownController  {
 
 	/**
 	 * updateBreakdown
+	 * 
 	 * @param id
 	 * @param breakdown
 	 * @param result
@@ -83,15 +85,15 @@ public class BreakdownController  {
 
 	/**
 	 * deleteBreakdown
+	 * 
 	 * @param id
 	 * @return
 	 */
 	@GetMapping("/deletebreakdown/{id}")
 	@Transactional
-	public String deleteBreakdown(@PathVariable("id") int id) {
+	public String deleteBreakdown(@PathVariable("id") Long id) {
 		breakdownService.delete(id);
 		return "redirect:/breakdowns";
 	}
-	
-	
+
 }
